requirements.txt
chromadb 
sentence-transformers
groq
python-dotenv